+++ socialhome/frontend/tests/components/streams/stamped_elements/ProfileStampedElement.tests.js
@@ -6,37 +6,32 @@ import Vue from "vue"
 import Vuex from "vuex"
 
 import ProfileStampedElement from "frontend/components/streams/stamped_elements/ProfileStampedElement.vue"
-import {newApplicationStore} from "frontend/stores/applicationStore"
+import {getStore} from "../../../fixtures/store.fixtures"
 
 
 Vue.use(BootstrapVue)
 Vue.use(Vuex)
 
 describe("ProfileStampedElement", () => {
+    let store
+
     beforeEach(() => {
+        store = getStore()
         window.context = getContext({currentBrowsingProfileId: 26})
     })
 
     describe("computed", () => {
         describe("displayName", () => {
             it("should return name if set", () => {
-                window.context.profile.name = "Karl Marx"
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.name = "Karl Marx"
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().displayName.should.eql("Karl Marx")
             })
 
             it("should return FID if name is unset", () => {
-                window.context.profile.name = undefined
-                window.context.profile.fid = "http://example.com/profile"
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.name = undefined
+                store.state.application.profile.fid = "http://example.com/profile"
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().displayName.should.eql("http://example.com/profile")
@@ -45,39 +40,27 @@ describe("ProfileStampedElement", () => {
 
         describe("showProfileButtons", () => {
             it("should be false if user is not authenticated", () => {
-                window.context.profile.id = 42
-                window.context.currentBrowsingProfileId = 42
-                window.context.isUserAuthenticated = false
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.id = 42
+                store.state.application.currentBrowsingProfileId = 42
+                store.state.application.isUserAuthenticated = false
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().showProfileButtons.should.be.false
             })
 
             it("should be false if user is not visiting own profile", () => {
-                window.context.profile.id = 42
-                window.context.currentBrowsingProfileId = 51
-                window.context.isUserAuthenticated = true
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.id = 42
+                store.state.application.currentBrowsingProfileId = 51
+                store.state.application.isUserAuthenticated = true
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().showProfileButtons.should.be.false
             })
 
             it("should be true if user is visiting own profile", () => {
-                window.context.profile.id = 42
-                window.context.currentBrowsingProfileId = 42
-                window.context.isUserAuthenticated = true
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.id = 42
+                store.state.application.currentBrowsingProfileId = 42
+                store.state.application.isUserAuthenticated = true
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().showProfileButtons.should.be.true
@@ -86,39 +69,27 @@ describe("ProfileStampedElement", () => {
 
         describe("showProfileReactionButtons", () => {
             it("should be false if user is not authenticated", () => {
-                window.context.profile.id = 42
-                window.context.currentBrowsingProfileId = 42
-                window.context.isUserAuthenticated = false
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.id = 42
+                store.state.application.currentBrowsingProfileId = 42
+                store.state.application.isUserAuthenticated = false
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().showProfileReactionButtons.should.be.false
             })
 
             it("should be true if user is not visiting own profile", () => {
-                window.context.profile.id = 42
-                window.context.currentBrowsingProfileId = 51
-                window.context.isUserAuthenticated = true
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.id = 42
+                store.state.application.currentBrowsingProfileId = 51
+                store.state.application.isUserAuthenticated = true
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().showProfileReactionButtons.should.be.true
             })
 
             it("should be false if user is visiting own profile", () => {
-                window.context.profile.id = 42
-                window.context.currentBrowsingProfileId = 42
-                window.context.isUserAuthenticated = true
-                const store = new Vuex.Store({
-                    state: {},
-                    modules: {applicationStore: newApplicationStore()}
-                })
+                store.state.application.profile.id = 42
+                store.state.application.currentBrowsingProfileId = 42
+                store.state.application.isUserAuthenticated = true
                 const target = mount(ProfileStampedElement, {store})
 
                 target.instance().showProfileReactionButtons.should.be.false